backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
python-multipart
pydub
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests